Refactor component structure and add new Profile, TestPage, LoadingSpinner, and Alert components; update Navbar and ProtectedRoute for improved routing
