+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Navbar from './components/layout/Navbar';
+import Navbar from './components/container/layout/Navbar';
 import Home from './pages/Home';
 import Products from './pages/Products';
 import Login from './pages/Login';
