+++ src/components/presentational/Profile.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { logout } from '../features/auth/authSlice';
-import { useAuth } from '../hooks/useAuth';
+import { logout } from '../../features/auth/authSlice';
+import { useAuth } from '../../hooks/useAuth';
 
 const Profile = () => {
   const dispatch = useDispatch();
